Struct linera_alloy_transport_http::Http

pub struct Http<T> { /* private fields */ }

An Http transport.

The user must provide an internal http client and a URL to which to connect. It implements Service<Box<RawValue>>, and therefore Transport.

Currently supported clients are:

reqwest

Implementations

impl Http<Client>

pub fn new(url: Url) -> Self

Create a new Http transport.

impl<T> Http<T>

pub const fn with_client(client: T, url: Url) -> Self

Create a new Http transport with a custom client.

pub fn set_url(&mut self, url: Url)

Set the URL.

pub fn set_client(&mut self, client: T)

Set the client.

pub fn guess_local(&self) -> bool

Guess whether the URL is local, based on the hostname.

The output of this function is best-efforts, and should be checked if possible. It simply returns true if the connection has no hostname, or the hostname is localhost or 127.0.0.1.

pub const fn client(&self) -> &T

Get a reference to the client.

pub fn url(&self) -> &str

Get a reference to the URL.